code: move json to parsers dir
