+++ code/header/parsers/json.h
@@ -1,4 +1,4 @@
-// file: header/json.h
+// file: header/parsers/json.h
 
 #ifdef ZPL_EDITOR
 #include <zpl.h>
