+++ code/source/parsers/json.c
@@ -1,4 +1,4 @@
-// file: source/json.c
+// file: source/parsers/json.c
 
 ////////////////////////////////////////////////////////////////
 //
